requirements.txt
Flask==2.3.3
Flask-SQLAlchemy==3.1.1
Flask-JWT-Extended==4.5.3
Flask-Cors==4.0.0
Flask-Mail==0.9.1
pymongo==4.6.1
python-dotenv==1.0.0
pyotp==2.9.0
Werkzeug==2.3.7
email-validator==2.1.0
gunicorn==21.2.0
mongoengine==0.27.0
passlib==1.7.4
pyjwt==2.8.0 